# baselines/openapi.yaml
openapi: 3.1.0
info:
  title: OReilly Remote POS API
  version: 1.0.0
  description: |
    Workshop hero API showing reference-based customer usage for operational flows.
  x-owner: store-platform
  x-domain: remote-pos
  x-workshop-pattern: reference-live-entity
  x-api-lifecycle: workshop-pilot
servers:
  - url: http://localhost:3000
tags:
  - name: remote-invoices
    description: Remote invoice operations that reference live customer ownership.
  - name: stock-transfers
    description: Stock transfer operations that reuse shared line item contract shapes.
paths:
  /remote-invoices:
    post:
      tags:
        - remote-invoices
      summary: Create a remote invoice
      operationId: createRemoteInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoteInvoiceCreateRequest'
      responses:
        '201':
          description: Remote invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoteInvoiceCreateResponse'
              examples:
                default:
                  value:
                    invoiceNumber: INV-1001
                    invoiceTotal: 41.98
                    transferResponses: []
                    customerValidationStatus: skipped
                    notificationStatus: skipped
        '400':
          $ref: '../shared-contracts/common-schemas.yaml#/components/responses/BadRequest'
  /remote-invoices/{invoiceNumber}:
    get:
      tags:
        - remote-invoices
      summary: Get a stored remote invoice
      operationId: getRemoteInvoice
      parameters:
        - name: invoiceNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Stored remote invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoteInvoice'
              examples:
                default:
                  value:
                    invoiceNumber: INV-1001
                    storeNumber: 221
                    counterNumber: 4
                    customer:
                      customerNumber: 90001234
                    delivery: true
                    items:
                      - line: IGN
                        item: PLUG-001
                        quantity: 2
                        price: 12.99
                    freight: 5
                    serviceCharges: 2.5
                    invoiceTotal: 33.48
                    transferResponses: []
                    customerValidationStatus: validated
                    notificationStatus: accepted
                    createdAt: '2026-04-10T12:00:00.000Z'
        '404':
          $ref: '../shared-contracts/common-schemas.yaml#/components/responses/NotFound'
  /stock-transfers:
    post:
      tags:
        - stock-transfers
      summary: Create a stock transfer
      operationId: createStockTransfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockTransferCreateRequest'
      responses:
        '201':
          description: Stock transfer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockTransferResponse'
        '400':
          $ref: '../shared-contracts/common-schemas.yaml#/components/responses/BadRequest'
  /stock-transfers/{transferId}:
    get:
      tags:
        - stock-transfers
      summary: Get a stock transfer
      operationId: getStockTransfer
      parameters:
        - name: transferId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Stock transfer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockTransferResponse'
        '404':
          $ref: '../shared-contracts/common-schemas.yaml#/components/responses/NotFound'
components:
  schemas:
    CustomerRef:
      type: object
      additionalProperties: false
      required:
        - customerNumber
      properties:
        customerNumber:
          type: integer
          format: int64
          minimum: 1
    LineItemQuantityWithPrice:
      type: object
      additionalProperties: false
      required:
        - line
        - item
        - quantity
        - price
      properties:
        line:
          type: string
        item:
          type: string
        quantity:
          type: integer
          format: int32
          minimum: 1
        price:
          type: number
          minimum: 0
        corePrice:
          type: number
          minimum: 0
    StockTransferStatus:
      type: string
      enum:
        - success
        - failure
        - pending
    StockTransferCreateRequest:
      type: object
      additionalProperties: false
      required:
        - counterNumber
        - receivingStoreNumber
        - customer
        - orderType
        - itemDetails
      properties:
        counterNumber:
          type: integer
          format: int32
          minimum: 1
        receivingStoreNumber:
          type: integer
          format: int32
          minimum: 1
        customer:
          $ref: '#/components/schemas/CustomerRef'
        orderType:
          type: string
        shipToName:
          type: string
        shipToAddress:
          type: string
        shipToCity:
          type: string
        shipToState:
          type: string
        itemDetails:
          type: array
          minItems: 1
          items:
            $ref: '../shared-contracts/common-schemas.yaml#/components/schemas/LineItemQuantity'
    StockTransferResponse:
      type: object
      additionalProperties: false
      required:
        - transferId
        - status
        - location
        - locationCityState
        - items
      properties:
        transferId:
          type: string
        status:
          $ref: '#/components/schemas/StockTransferStatus'
        location:
          type: string
        locationCityState:
          type: string
        items:
          type: array
          items:
            $ref: '../shared-contracts/common-schemas.yaml#/components/schemas/LineItemQuantity'
    RemoteInvoiceCreateRequest:
      type: object
      additionalProperties: false
      required:
        - storeNumber
        - counterNumber
        - customer
        - delivery
        - items
      properties:
        storeNumber:
          type: integer
          format: int32
          minimum: 1
        counterNumber:
          type: integer
          format: int32
          minimum: 1
        customer:
          $ref: '#/components/schemas/CustomerRef'
        delivery:
          type: boolean
        shippingInstructions:
          type: string
        shipToName:
          type: string
        shipToAddress:
          type: string
        shipToCity:
          type: string
        shipToState:
          type: string
        poNumber:
          type: string
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/LineItemQuantityWithPrice'
        freight:
          type: number
          minimum: 0
        serviceCharges:
          type: number
          minimum: 0
        stockTransfers:
          type: array
          items:
            $ref: '#/components/schemas/StockTransferResponse'
    RemoteInvoiceCreateResponse:
      type: object
      additionalProperties: false
      required:
        - invoiceNumber
        - invoiceTotal
        - transferResponses
        - customerValidationStatus
        - notificationStatus
      properties:
        invoiceNumber:
          type: string
        invoiceTotal:
          type: number
        transferResponses:
          type: array
          items:
            $ref: '#/components/schemas/StockTransferResponse'
        customerValidationStatus:
          type: string
          enum:
            - skipped
            - validated
            - not-found
        notificationStatus:
          type: string
          enum:
            - skipped
            - accepted
            - duplicate
            - failed
    RemoteInvoice:
      type: object
      additionalProperties: false
      required:
        - invoiceNumber
        - storeNumber
        - counterNumber
        - customer
        - delivery
        - items
        - invoiceTotal
        - transferResponses
        - customerValidationStatus
        - notificationStatus
        - createdAt
      properties:
        invoiceNumber:
          type: string
        storeNumber:
          type: integer
          format: int32
        counterNumber:
          type: integer
          format: int32
        customer:
          $ref: '#/components/schemas/CustomerRef'
        delivery:
          type: boolean
        shippingInstructions:
          type: string
        shipToName:
          type: string
        shipToAddress:
          type: string
        shipToCity:
          type: string
        shipToState:
          type: string
        poNumber:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/LineItemQuantityWithPrice'
        freight:
          type: number
        serviceCharges:
          type: number
        invoiceTotal:
          type: number
        transferResponses:
          type: array
          items:
            $ref: '#/components/schemas/StockTransferResponse'
        customerValidationStatus:
          type: string
          enum:
            - skipped
            - validated
            - not-found
        notificationStatus:
          type: string
          enum:
            - skipped
            - accepted
            - duplicate
            - failed
        createdAt:
          type: string
          format: date-time